+++ src/pch.h
@@ -8,8 +8,8 @@
 #define PCH_H
 
 #define SADX_AP_VERSION_MAJOR 0
-#define SADX_AP_VERSION_MINOR 0
-#define SADX_AP_VERSION_PATCH 5
+#define SADX_AP_VERSION_MINOR 1
+#define SADX_AP_VERSION_PATCH 0
 
 // add headers that you want to pre-compile here
 #include "framework.h"