+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - Re-enabled mission/trial options on the menu
 - Last Story takes the player directly to the Perfect Chaos fight
+- The demo cutscenes no longer give checks
 
 ## [v0.0.5] - 2024-07-07