+++ CHANGELOG.md
@@ -5,12 +5,17 @@ All notable changes to this project will be documented in this file.
 ## [v0.0.6] - ?
 
 ### Changed
+
 - Re-enabled mission/trial options on the menu.
 - Last Story takes the player directly to the Perfect Chaos fight.
 - The demo cutscenes no longer give checks.
 - Unlocks are shown on the character select screen and pause menu.
-  - It only shows the characters that are playable in the randomizer.
+    - It only shows the characters that are playable in the randomizer.
 - Prevent Supersonic for being selected early.
+
+### Added
+
+- Added message to check your configs if the connection takes too long.
 
 ## [v0.0.5] - 2024-07-07